+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.0.2
+* Adding optional callback on stream error
+    * Also adding when response status code is >= 400
+
 ### 1.0.1
 * Adding streaming progress when no `content-length` header is set